Use shell built-in case to determine toolkit version
In Dockerfile_binary_openvino, the use of pwd piped to grep and checking
grep's result code is here replaced with a simple shell native case
instead. This increases readability and maintainability; should there be
difference between toolkit version it is quite simple to extend the
case. (A bonus is increased performance by using shell natives, even if
that is hardly noticable on a modern computer.)

Change-Id: I40ad3ac1e98c45e8b197498130412ebe8f03705a
Signed-off-by: Joakim Roubert <joakimr@axis.com>