Jenkins: virtualenv in separate step for Jenkins worker
If not separated on purpose, functional tests run in docker would use
the same virtualenv directory as Jenkins workspace. This leads to errors
due to different path in python virtual environment.

As a different approach we could mount the OVMS directory in the same
path as on host. Different approaches still needs to be investigated.

Signed-off-by: Krzysztof Romanowski <krzysztof.romanowski@intel.com>