Limit available nireq values for model instance
Issue was that both very high and 0 does not make sense for that
parameter since if you provide 0 there is no sense in providing model
and if you provide too high there is no gain since there is limit to
which inferences can be parametrized.

For now we set arbitrary value, will set to metrics got from OpenVINO in
next release.

JIRA:CVS-36712, CVS-38350, CVS-32899