Add gather node implementation (#480)
This includes GatherNodeInputHandler. Added
NodeSessionMetadata::getShardId for each subsession shard identification
in case of mutliple levels of session collapse.

Additionally:
* Remove Node::setInputs(dependency, BlobMap) as it became unnecessary
* Fix problem when no release on session resources was called
Added test for that case as well. It was working by accident since in
Node::fetchResults() we erased session from which we got. However we
didn't do it with the release API call.
* Additionally remove unused code after Node & NodeSession integration

JIRA:CVS-46710