Increase wait time for dag loaded state
Without this we sporadically hit the PIPELINE_NOT_LOADED_YET status
in CI. This can hapen under heavy machine workload as the timeout for
waiting was 10ms and sporadically it was ~70ms on some machines.

Adding this status to allowed statuses list for this test would lower
the test value since we can hit it when eg. there is unsuccessful
validation.

If the threshold is not increased the PIPELINE_NOT_LOADED_YET status
code would encourage to try to send request again so it seems that's
better to wait a bit longer than force client to retry with request.

JIRA:CVS-63360