Improved OVMS Mediapipe calculators logging and error handling
* Remove doubled filename & line number in log lines
* Catch possible C-API infer errors inside OVMS Calculators
While those exceptions are catched at higher level, we can log
additional information in lower stack. Final error handling is yet to be
implemented together with C++ Model API
* Added warning about double loggers configuring for easier debug
* OVMS will set Mediapipe log level to INFO when OVMS is in DEBUG/TRACE
This is due to Glog not having compiled in debug when in build is not
debug.

Note:
Ideally we could format Glog logs in the same way as SPDLOG however this
requires glog to be comiled with flag enabling custom prefix
GLOG_CUSTOM_PREFIX_SUPPORT and forcing this flag through MP.
Calling InitGoogleLogging for some reason breaks appropiate log level
setting so it is skipped now.

JIRA:CVS-109174