Fix sporadic timeout issue in test
Issue was that on particularly slow or overloaded machine the config
reload could not happen within 1.8s window. Threshold was increased to
5s but at the same time test utility function to check for config reload
exits as soon as the config is already loaded.

ModelManager::configFileReloadNeeded method is tested separately.

Additionally for test setups watcher check interval is set to lower
value.

JIRA:CVS-109709