Stop build if cuda is not complete
Sometimes cudnn fails to download. This causes the transaction to fail,
however, yum/dnf both return 0. This means that the build continues until
it fails to build the cuda wheel which is much later in the build. With
this patch applied, it stops the build if everything did not download.