Switch Dockerfile.redhat over to use dnf
There is a mix of dnf and yum being used throughout the Dockerfile.
This switches it over to only use dnf. As a result, yum-utils is not
needed. This also adds more parallelism in downloads and more retries.
The nvidia repo sometimes breaks connections. Hopefully the extra
retries lets us be successful in setting up the build environment.