Rtinfo (#2736)
Include in REST KServe API for metadata call the content of RT_info. There is extra json object including the AnyMap structure with the contnet of get_rt_info included in the model.
The change affect also CAPI interface. The ModelMetadata response include now the RTInfo map adjusted in such a way that values with ov::META are replaced with ov:AnyMap object.

---------

Co-authored-by: Dariusz Trawinski <dariusz.trainski@intel.com>