Check graph pool
TODO:
-> this requires additional patch in MP to reset initialized_ flag in
CalculatorGraph and verify if that works. Previous MP tests with reruns
worked due to using AddVectorSink which changes the underlying graph and
does not use OutputStreamPollers. Need to verify if change in MP will
enable graph pool or we need to go back to thread pool.

Rebased. Tests build.