Allow IPv6 addresses
The REST API works with IPv6 addresses. But the address validation only
supports IPv4. gRPC on the otherhand is supposed to work with IPv6, but
has some kind of problem with the current version of gPRC. It does
support IN6ADDR_ANY, though.

The origin of the regex is here:
https://stackoverflow.com/questions/53497/regular-expression-that-matches-valid-ipv6-addresses