Fix running tests for Docker builds
If you build with a Dockerfile and pass --build-arg RUN_TESTS=1,
it will fail to run the tests because the prepare_llm_models.sh
calls a model_export python script at a specific directory location.
This patch copies model_export.py into the image and places it at
the expected location.