Install openvino-tokenizer python module in image
I was working through a demo where it said to run
"pip install -U optimum[diffusers,openvino]".
When I did that, I noticed that it downloaded:
openvino_tokenizers-2025.1.0.0-py3-none-manylinux2014_x86_64.whl.
However, the tokenizer is built and installed in the image
by the Dockerfile. After more digging, I found that openvino's
python module is not installed in the image. This patch installs
the openvino-tokenizer python module in the image and uses a
symlink to point to the tokenizer shared object to save space.