Allow IPv6 addresses (#3295)
* Allow IPv6 addresses
The REST API works with IPv6 addresses. But the address validation only
supports IPv4. gRPC on the otherhand is supposed to work with IPv6, but
has some kind of problem with the current version of gPRC. It does
support IN6ADDR_ANY, though.

The origin of the regex is here:
https://stackoverflow.com/questions/53497/regular-expression-that-matches-valid-ipv6-addresses

Add IPv6 unit tests
This tests various kinds of IPv6 addresses. It should be noted that
Link-local IPv6 with zone index address types are not supported. They
are tested to ensure they are rejected.

* style check