ImageGen: Static Shape, Multi Device (#3440)
CVS-168932

Support for Static Shape operation so that workloads can run on NPU (but allowed also for other devices)
Support for device selection for different pipeline stages (text encode/denoise/vae) so that issues on NPU can be mitigated by picking different device for faulty part of pipeline (for example SD v1.5 512x512 on MTL - there is ticket for that)

This enables NPU from perspective of OVMS, however fix on GenAI is still required. It is present on 2025/2 release branch of GenAI, and master.