Fix opencl-header installation
The code "as is" installs opencl-headers from rhel8 because the rhel9 one
leads to compilation failures. However, it immediately does a dnf -y update
which upgrades the opencl package to the rhel 9 version that was being
avoided. Compilation failures follow at 64% of the way through openvino's
code.

The fix is to get rid of the dnf -y update because the image was already
updated back at the first use of dnf.